Restructured my modules so that state was by itself in database.js, I also change the original load screen to contain a blank slate of orders
