+++ scripts/JewelryStyles.js
@@ -1,4 +1,4 @@
-import { getOrderBuilder, getStyles, setStyle } from "./database.js"
+import { getOrderBuilder, getStyles, setStyle } from "./dataAccess.js"
 
 
 document.addEventListener(
