+++ scripts/Metals.js
@@ -1,4 +1,4 @@
-import { getMetals, getOrderBuilder, setMetal } from "./database.js"
+import { getMetals, getOrderBuilder, setMetal } from "./dataAccess.js"
 
 
 document.addEventListener(
